server/requirements.txt
python-socketio
gunicorn
uvicorn[standard]